pyproject.toml
[project]
name = "autobots-agent-jarvis"
version = "0.1.0"
description = "Jarvis - Multi-agent AI Assistant Demo"
readme = "README.md"
authors = [
    {name = "PK", email = "pro.kamath@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "autobots-devtools-shared-lib==0.1.4a5",
    "chainlit>=2.9.6",
    "langchain>=1.0.0",
    "langchain-google-genai>=4.2.0",
    "langfuse>=3.12.1",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.5.1",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
]

[tool.poetry]
name = "autobots-agent-jarvis"
version = "0.1.0"
description = "Jarvis - Multi-agent AI Assistant Demo"
authors = ["PK <pro.kamath@gmail.com>"]
packages = [
    { include = "autobots_agents_jarvis", from = "src" }
]

[tool.poetry.dependencies]
# Local path dependency (Poetry-specific)
autobots-devtools-shared-lib = {path = "../autobots-devtools-shared-lib", develop = true}


[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "S",      # flake8-bandit (security)
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RET",    # flake8-return
    "TRY",    # tryceratops
    "PERF",   # perflint
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "S101",   # Allow assert in tests
    "TRY003", # avoid specifying long messages outside exception class
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S106", "ARG001", "ARG002", "PLR2004"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--strict-config",
    "--cov=src/autobots_agents_jarvis",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-branch",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
include = ["src", "tests"]
exclude = ["**/__pycache__", ".venv"]
venvPath = ".."
venv = ".venv"
extraPaths = ["src", "../autobots-devtools-shared-lib/src"]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
